fix(sim): all 8 firmware survival tests pass — remove debug instrumentation
- NVIC bit computation fixed (3 bus/mod.rs locations)
- pending_exceptions widened u32→u64 to handle extended IRQ ranges
- Infinite ISR re-entry fixed via active_exception tracking + ISPR clear on take
- DWT peripheral size expanded 0x40→0x1000 to avoid MemoryViolation in dwt_init
- Bit-band aliases implemented in bus write_u32/read_u32 for PLLON writes
- Interrupt-driven UART TX chain fully functional (TXEIE→IRQ→ISR→DR drain)
- Remove all temporary debug eprintln! instrumentation added during investigation

+++ crates/core/src/snapshot.rs
@@ -27,7 +27,7 @@ pub struct ArmCpuSnapshot {
     pub pc: u32,
     pub xpsr: u32,
     pub primask: bool,
-    pub pending_exceptions: u32,
+    pub pending_exceptions: u64,
     pub vtor: u32,
 }
